+++ Assets/Scripts/Unity/Graphics/SpriteCache.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using UnityEngine;
 using Ventura.GameLogic;
-using Ventura.Loaders;
 
 namespace Ventura.Unity.Graphics
 {
@@ -11,20 +10,46 @@ namespace Ventura.Unity.Graphics
         private static SpriteCache _instance = new SpriteCache();
         public static SpriteCache Instance { get { return _instance; } }
 
-        private Dictionary<Entity, Sprite> _sprites = new();
+        private Dictionary<string, Sprite> _sprites = new();
 
-        public Sprite GetSprite(Entity e)
+        //public Sprite GetSprite(Entity e)
+        //{
+        //    Sprite sprite;
+        //    if (_sprites.ContainsKey(e))
+        //        sprite = _sprites[e];
+        //    else
+        //    {
+        //        sprite = SpriteLoader.Load(e);
+        //        _sprites.Add(e, sprite);
+        //    }
+
+        //    return sprite;
+        //}
+
+        public Sprite GetSprite(string spriteId)
         {
             Sprite sprite;
-            if (_sprites.ContainsKey(e))
-                sprite = _sprites[e];
+            if (_sprites.ContainsKey(spriteId))
+                sprite = _sprites[spriteId];
             else
             {
-                sprite = SpriteLoader.Load(e);
-                _sprites.Add(e, sprite);
+                sprite = LoadSprite(spriteId);
+                _sprites.Add(spriteId, sprite);
             }
 
             return sprite;
         }
+
+
+        private static Sprite LoadSprite(string spriteId)
+        {
+            var spriteName = GraphicsConfig.SpriteFiles[spriteId];
+            var sprite = Resources.Load<Sprite>($"Sprites/{spriteName}");
+            if (sprite == null)
+                throw new GameException($"!! Sprite not found: {spriteName}");
+
+            return sprite;
+        }
+
     }
 }